build releases automatically when pushing tags
This will add a new tag and build a release on github if you
push any annotated tag to the upstream repository.
Additionally all commits since the last tag will be added to
the release notes for further editing.

Lastly, we move the ccache directory to inside output/ in
order to cache it as well. The downside obviously is that
you'll loose the ccache history when you clean your build
directory.