multi-gadget: wait for udev settle before setting up gadget
The script depends on the state of attached hardware, but didn't do
udevadm settle until the end, this could lead to a race condition where
the camera doesn't get detected early enough during boot.

Make the service wait for systemd-udev-settle.service before starting,
so we can be reasonably sure that if a camera is attached it will be
detected.