Enable only controls that the RPi camera supports
The values that are currently set by default in the patch for
updating the f_uvc.c kernel driver file will try to enable some
controls that are not supported by the 8 Mpixel (v2) camera on
the RaspberryPi 4B HW and leaves out some that are supported.

Fix the values to match exactly the controls that are possible.

Signed-off-by: Soren Friis <soren.friis@intel.com>