Add Raspberry Zero-2 support
This change adds the following functionalities:

- Add Raspberry Pi Zero 2W support
- Consolidate Raspberry Zero support to one image
- Add the ability to override buildroot packages (in this case,
we override rpi-firmware package due to buildroot upstream hasn't
added support for Raspberry Pi Zero 2W)
- Allow build-showmewebcam.sh to take optional parameters

The build-showmewebcam.sh can now be called with a optional buildroot
package like so:

./build-showmewebcam.sh raspberrypi0 rpi-firmware
./build-showmewebcam.sh raspberrypi0 menuconfig

And so on. This simplifies manually calling those commands when we
just try to build a single package.

Fixes #168