configs: Use correct defconfigs
It turns out that maintaining our linux configs is not a good idea.
This makes adding new board supports very hard, while we can mirror
what is buildroot does for the pis: Just use the foundation's
defconfig.

So this change deletes all of our linux defconfigs and just add
the modules that we want for showmewebcam to function, namely
board/linux-additional-modules.config.

This also add the post-image changes that is needed for this
almost-stock kernel to boot quickly and with the modules we desire.